+++ tests/test_util.py
@@ -28,6 +28,26 @@ class TestUtils(unittest2.TestCase):
         self.assertEqual(searchargs((['spam in(1, 2)'],)),
                          ([('spam', 'in', (1, 2))],))
 
+        # Standard comparison operators
+        self.assertEqual(searchargs((['ham=2'],)), ([('ham', '=', 2)],))
+        self.assertEqual(searchargs((['ham!=2'],)), ([('ham', '!=', 2)],))
+        self.assertEqual(searchargs((['ham>2'],)), ([('ham', '>', 2)],))
+        self.assertEqual(searchargs((['ham>=2'],)), ([('ham', '>=', 2)],))
+        self.assertEqual(searchargs((['ham<2'],)), ([('ham', '<', 2)],))
+        self.assertEqual(searchargs((['ham<=2'],)), ([('ham', '<=', 2)],))
+
+        # Operators rarely used
+        self.assertEqual(searchargs((['status =like Running'],)),
+                         ([('status', '=like', 'Running')],))
+        self.assertEqual(searchargs((['status=like Running'],)),
+                         ([('status', '=like', 'Running')],))
+        self.assertEqual(searchargs((['status =ilike Running'],)),
+                         ([('status', '=ilike', 'Running')],))
+        self.assertEqual(searchargs((['status =? Running'],)),
+                         ([('status', '=?', 'Running')],))
+        self.assertEqual(searchargs((['status=?Running'],)),
+                         ([('status', '=?', 'Running')],))
+
         self.assertRaises(ValueError, searchargs, (['spam.hamin(1, 2)'],))
         self.assertRaises(ValueError, searchargs, (['spam.hamin (1, 2)'],))
         self.assertRaises(ValueError, searchargs, (['spamin (1, 2)'],))